# doc/swagger/AoL-Teacher.yaml
openapi: 3.0.0
info:
  title: AoL Teacher Interface
  description: |-
    AoL Teacher API specification.

  contact:
    email: sudhagar@abovecloud9.ai
  version: 0.0.4
...
servers:
  - url: https://api.abovecloud9.ai/api/v1 # placeholder
...
paths:
...
  /products:
    get:
      tags:
        - product
      summary: List all products
      description: Returns a list of all available products
      operationId: listProducts
      parameters:
        - name: q
          in: query
          description: Search string
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Page number (for pagination)
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          description: Number of items in the response
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearer: []
    
    post:
      tags:
        - product
      summary: Create a new product
      description: Add a new product to the system
      operationId: createProduct
      requestBody:
        description: Product object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseProduct'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearer: []
  /products/{id}:
    get:
      tags:
        - product
      summary: Get product by ID
      description: Returns a single product by its ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          description: ID of the product to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearer: []
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: integer
              format: int32
            message:
              type: string
            details:
              type: string
            traceId:
              type: string
    BaseProduct:
      type: object
      required:
        - name
        - title
        - ctype
        - baseProductId
      properties:
        # Just title is enough. This is an internal field. Commented until integration testing.
        # extName:
        #   type: string
        #   example: "Art of Living Course - Part 1"
        title:
          type: string
          example: "Art of Living Course Part 1 - In person"
        ctype:
          type: string
          example: "C-12345"
        # Note: This is not required. Commented until integration testing.
        # baseProductExtId:
        #   type: string
        #   example: "54321"
        durationDays:
          type: number
          format: int
          example: 3
        visibility:
          type: string
          nullable: true
          enum: [Public, Unlisted, null]
          example: "Public"
        maxAttendees:
          type: number
          format: int
          example: 100
        format:
          type: string
          enum: [In Person, Online, Destination Retreats]
          example: "In Person"
    Product:
      allOf: # Combines the BaseProduct and the inline model
        - $ref: "#/components/schemas/BaseProduct"
        - type: object
          required:
            - rootCause
          properties:
            id:
              type: integer
              format: int64
              example: 10000000
...
  securitySchemes:
    bearer:
      type: apiKey
      name: Bearer
      in: header
    sample_auth_unused:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://abovecloud9.ai/oauth/authorize
          scopes:
            write:courses: modify courses
            read:courses: read courses